+++ NEWS.md
@@ -1,5 +1,7 @@
 # genieBPC (development version)
 
+* Updated `genieBPC::nsclc_test_data` to ensure unique sample ids for CNA file.
+
 # genieBPC (1.1.0.9000)
 
 * Updated `genieBPC::nsclc_test_data` to randomly sample from cancer diagnosis dataset, stratified by stage, and incorporate genomic data into dataset.